refactor: 이미지 수정 분리
